+++ src/libraries/clients_kernel/Architecture_ABC.h
@@ -35,7 +35,7 @@ public:
 
     //! @name Operations
     //@{
-    virtual void Initialize( float height, unsigned int floorNumber, unsigned int parkingFloors, const std::string& roofShape,
+    virtual void Initialize( unsigned int height, unsigned int floorNumber, unsigned int parkingFloors, const std::string& roofShape,
                              const std::string& material, float occupation, float trafficability ) = 0;
     virtual const std::string& GetMaterial() const = 0;
     virtual unsigned int GetFloorNumber() const = 0;
